# documents/edificad-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Edificad API",
    "description": "Espeicificação da API Edificad, voltada para a gestão da ONG Edificando Vidas. ",
    "version": "0.0.1",
    "contact": {
      "name": "Repositório",
      "url": "https://github.com/jarnunes/edificad"
    },
    "license": {
      "name": "GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    }
  },
  "servers": [
    {
      "url": "https://edificad-production.up.railway.app/api",
      "description": "Produção"
    },
    {
      "url": "http://edificad.com:5000/api",
      "description": "Desenvolvimento"
    }
  ],
  "paths": {
...
    "/distribuicao-cesta": {
      "get": {
        "tags": [
          "Distribuição Cesta"
        ],
        "summary": "Retorna um array com todas as distribuições registradas",
        "parameters": [
          {
            "name": "cesta",
            "in": "query",
            "description": "Nome da cesta",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cpfBeneficiario",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cpfVoluntario",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "data",
            "in": "query",
            "description": "Data de distribuição da cesta",
            "schema": {
              "type": "string",
              "example": "2023-10-28"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribuicaoCestaArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Distribuição Cesta"
        ],
        "summary": "Registar nova distribuição de cesta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistribuicaoCestaPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribuicaoCesta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "auth": []
    }
  ],
  "components": {
    "schemas": {
...
      "BeneficiarioDistribuicaoCesta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "nome": {
            "type": "string",
            "example": "Beneficiário A"
          },
          "email": {
            "type": "string",
            "example": "beneficiario@email.com"
          },
          "cpf": {
            "type": "string",
            "example": "000.000.000-00"
          },
          "telefone": {
            "type": "string",
            "example": "(99) 99999-9999"
          },
          "data_nascimento": {
            "type": "string",
            "example": "yyyy-mm-dd"
          },
          "qtde_projetos_participado": {
            "type": "integer",
            "example": 5
          }
        }
      },
...
      "CestaDistribuicao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "nome": {
            "type": "string",
            "example": "CESTA 01"
          }
        }
      },
...
      "VoluntarioDistribuicaoCesta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "nome": {
            "type": "string",
            "example": "Voluntario A"
          },
          "email": {
            "type": "string",
            "example": "voluntario@email.com"
          },
          "cpf": {
            "type": "string",
            "example": "000.000.000-00"
          },
          "telefone": {
            "type": "string",
            "example": "(99) 99999-9999"
          },
          "data_nascimento": {
            "type": "string",
            "example": "yyyy-mm-dd"
          },
          "qtde_projetos_participado": {
            "type": "integer",
            "example": 5
          }
        }
      },
...
      "DistribuicaoCesta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "data_hora": {
            "type": "string",
            "example": "2023-10-27T20:36:38.150664"
          },
          "beneficiario": {
            "$ref": "#/components/schemas/BeneficiarioDistribuicaoCesta"
          },
          "voluntario": {
            "$ref": "#/components/schemas/VoluntarioDistribuicaoCesta"
          },
          "cesta": {
            "$ref": "#/components/schemas/CestaDistribuicao"
          }
        }
      },
      "DistribuicaoCestaArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/DistribuicaoCesta"
        }
      },
      "DistribuicaoCestaPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "data_hora": {
            "type": "string",
            "example": "2023-10-27T20:36:38.150664"
          },
          "beneficiario": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 21
              }
            }
          },
          "voluntario": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 244
              }
            }
          },
          "cesta": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 233
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}